Add `Builder` abstraction to simplify attestation.
So far, the code would distinguish between the attester interface and
the auxiliary information that's necessary for requesting and
attestation document.  This made things somewhat cumbersome; especially
when some auxiliary information (the enclave hashes) needs to be set
early whereas other information needs to be set at attestation time (the
nonce).

This PR adds a new abstraction, the `Builder` type, which bundles an
attester with a mechanism to set auxiliary information at initialization
time and at attestation time.