Attest request only if nonce is given.
So far, the config handler would always try to extract the nonce from
the request in order to attest the response.  That's often annoying when
you simply want to see the config and don't care about the attestation.

This PR makes attestation optional: if the nonce is given, the handler
is going to create an attestation; if not, there won't be.